pyproject.toml
[project]
name = "lazybrother"
version = "0.1.0"
description = "Backend service for candlestick chart analysis with investment advice"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pandas>=2.2.0",
    "pandas-ta>=0.3.14b1",
    "python-binance>=1.0.19",
    "yfinance>=0.2.40",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    "anthropic>=0.39.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
